% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Combine}
\alias{Combine}
\title{Example output from `KDE_combine_areas_fun()`}
\format{
...
\describe{
  \item{list}{}

}
}
\usage{
Combine
}
\description{
Created and used in the vignettes
}
\concept{Data}
\keyword{datasets}
