% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contours_all}
\alias{contours_all}
\title{Data set with all contours of the example tracks}
\format{
...
\describe{
  \item{species}{}
  \item{colony}{}
  \item{individ_id}{}
  \item{KDE_contour}{}
  \item{month}{}
  \item{individ_Year}{}
  \item{area_km2}{}
  \item{geometry}{}
  \item{W_Year}{}
  \item{retrieval_year}{}
  \item{h_par}{}
}
}
\usage{
contours_all
}
\description{
Created and used in the vignettes
}
\concept{Data}
\keyword{datasets}
