% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_tracks}
\alias{example_tracks}
\title{Individual example tracks}
\format{
Each row contains a location estimated linked to a specific date-time
\describe{
  \item{session_id}{unique identifier for the recorded track}
  \item{individ_id}{unique identifier for the tracked individual}
  \item{species}{describes which species was tracked}
  \item{colony}{describes which breeding colony the tracked individual is associated with}
  \item{longitude}{longitude information of the location estimate}
  \item{latitude}{latitude information of the location estimate}
  \item{sex}{sex information of the tracked individual}
  \item{Year}{The year associated with the date-time of the location estimate, determined from the date-time information}
  \item{month}{The month associated with the date-time of the location estimate, determined from the date-time information}
  \item{W_Year}{The winter (e.g. 2010/2011) associated with the location estimate. January - June relate to the previous winter and July - December relate to the upcoming winter (assuming breeding happens in the northern hemisphere)}
  \item{individ_year}{A unique identifier for the tracked individual in a specific winter}
  \item{retrieval_year}{}
}
}
\usage{
example_tracks
}
\description{
A data frame containing tracks from several individuals
}
\keyword{datasets}
