% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE_combine_areas_fun.R
\name{KDE_combine_areas_fun}
\alias{KDE_combine_areas_fun}
\title{Function to iteratively combine individual-level kernels to estimate within-colony variation}
\usage{
KDE_combine_areas_fun(vertices_sf, tot_loc_data, n_iterations = 20)
}
\arguments{
\item{vertices_sf}{An sf-object with individual-level kernels already filtered for a desired KDE contour percentage}

\item{tot_loc_data}{A dataframe with species-colony location information}

\item{n_iterations}{Set how many iterations you'd like to use}
}
\value{
A list containing:
  \describe{
    \item{Kernel_areas}{A tibble with KDE bootstrap results}
    \item{colony_projection}{The LAEA projection used}
  }
}
\description{
Function to iteratively combine individual-level kernels to estimate within-colony variation
}
