% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE_filter_tracks_fun.R
\name{KDE_filter_tracks_fun}
\alias{KDE_filter_tracks_fun}
\title{Filter for tracks with sufficient location estimates}
\usage{
KDE_filter_tracks_fun(
  dataset,
  month. = 12,
  min_points_per_track = 10,
  min_Ninds_per_col = 5
)
}
\arguments{
\item{dataset}{A tibble with location estimates from a tracked colony. Must contain columns with species andcolony information.
Also, a month-column should be present, indicating in which month(numeric) a location estimate was made. As well as a column with a unique identifier per yearly migration track (`individ_Year`).}

\item{month.}{The month for which you wish to determine which tracks have sufficient location estimates (numeric)}

\item{min_points_per_track}{Set to the minimum number of location estimates per month you deem sufficient in your analysis (numeric)}

\item{min_Ninds_per_col}{Set to the minimum number of individuals with sufficient migration tracks that you deem appropriate for you analysis (numeric)}
}
\value{
Returns a list with two elements:
1) a tibble containing all the migratory tracks of the colony that meet the required number of locations (as set in `min_points_per_track`);
2) a vector with the names of the tracks that meet the requirements.
}
\description{
`KDE_filter_tracks_fun()` takes a tibble with (GLS-)tracking data from a colony and determines which migratory tracks (within a specified month) contain sufficient location estimates.
It then goes on to check if a colony has enough tracked individuals with sufficient location estimates.
}
\concept{Function}
