% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE_Repr_per_individual_fun.R
\name{KDE_repr_per_individual_fun}
\alias{KDE_repr_per_individual_fun}
\title{Function to determine a saturation curve per number of included individuals.}
\usage{
KDE_repr_per_individual_fun(combined_areas_df)
}
\arguments{
\item{combined_areas_df}{An tibble with combined-area information as produced by `KDE_combine_areas_fun()`.}
}
\value{
A tibble containing estimated representativeness after including data from a given retrieval year, thus assessing all data collected up to that point.
}
\description{
Function to determine a saturation curve per number of included individuals.
}
