% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE_Repr_per_retrieval_year_fun.R
\name{KDE_repr_per_retrieval_year_fun}
\alias{KDE_repr_per_retrieval_year_fun}
\title{Function to determine representativeness of data collected up to and including certain retrieval years.}
\usage{
KDE_repr_per_retrieval_year_fun(
  contours_sf,
  colony_projection,
  n_iterations = 20
)
}
\arguments{
\item{contours_sf}{An sf-object with individual-level kernels for a single species-colony already filtered for a desired KDE contour percentage. contours_sf should also contain a column indicating retrieval year}

\item{colony_projection}{Provide a projection that is appropriate for the whole colony}

\item{n_iterations}{How many iterations should be used in the area-combining process}
}
\value{
A tibble containing estimated representativeness after including data from a given retrieval year, thus assessing all data collected up to that point.
}
\description{
Function to determine representativeness of data collected up to and including certain retrieval years.
}
