% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE_saturation_curve_fun.R
\name{KDE_saturation_curve}
\alias{KDE_saturation_curve}
\title{Fit saturation curve model to combined areas}
\usage{
KDE_saturation_curve(Kernel_areas)
}
\arguments{
\item{Kernel_areas}{Data frame with iterated combined areas}
}
\value{
A list with:
  \describe{
    \item{Kernel_areas}{Input data with A_mean and B_mean added}
    \item{Kernel_areas_summarised}{Summary table with fitted curve statistics}
    }
}
\description{
Fit saturation curve model to combined areas
}
