% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mean_smooths}
\alias{mean_smooths}
\title{Species specific smoothing parameters}
\format{
...
\describe{
  \item{species}{smoothing parameters (h_par) are provided for 15 seabird species}
  \item{mean_h}{mean values from track-specific smoothing parameters determined using LSCV}
}
}
\usage{
mean_smooths
}
\description{
A data frame containing species-mean smoothing parameters, calculated from smoothing parameters determined using LSCV per winter track
}
\keyword{datasets}
