% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retrieval_year_repr}
\alias{retrieval_year_repr}
\title{Example output from `KDE_Repr_per_retrieval_year()`}
\format{
...
\describe{
  \item{XXX}{}

}
}
\usage{
retrieval_year_repr
}
\description{
Used in the vignettes
}
\keyword{datasets}
