% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saturation_curve}
\alias{saturation_curve}
\title{Example data as would be returned from `KDE_saturation_curve_fun()`}
\format{
...
\describe{
  \item{species}{}
  \item{colony}{}
  \item{month}{}
  \item{n_inds}{}
  \item{it}{}
  \item{perc}{}
  \item{area}{}
  \item{h_par}{}
  \item{A_mean}{}
  \item{B_mean}{}
  \item{Repr}{}

}
}
\usage{
saturation_curve
}
\description{
A data frame containing information on combined contours surface areas and the parameters of the Michaelis-Menten curve fitted to this data.
}
\concept{Data}
\keyword{datasets}
