% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sens_MM}
\alias{Sens_MM}
\title{Example data to assess uncertainty of the estimated colony representativeness}
\format{
...
\describe{
  \item{species}{}
  \item{colony}{}
  \item{N_ind}{}
  \item{A_mean}{}
  \item{B_mean}{}
  \item{Repr_5}{}
  \item{perc}{}
}
}
\usage{
Sens_MM
}
\description{
A data frame example data that will help indicate the uncertainty around the estimated colony representativeness
}
\keyword{datasets}
