% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_curves_tot_PN}
\alias{species_curves_tot_PN}
\title{Mutli-species data set for a grouped analysis}
\format{
...
\describe{
  \item{species}{}
  \item{colony}{}
  \item{N_inds}{}
  \item{KDE_contour}{}
  \item{Repr}{}
  \item{Desired_var_threshold}{}
  \item{N_inds_desired_perc}{}
  \item{A_mean}{}
  \item{B_mean}{}
  \item{Perc_desired}{}
  \item{PN_prop_missing}{}
}
}
\usage{
species_curves_tot_PN
}
\description{
A data frame containing many species-colonies and their estimated colony representativeness
}
\keyword{datasets}
