% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE_filter_tracks_fun.R
\name{KDE_filter_tracks_fun}
\alias{KDE_filter_tracks_fun}
\title{Filter for tracks to be assessed in the colony representativeness analysis}
\usage{
KDE_filter_tracks_fun(
  dataset,
  species.,
  colony.,
  month. = 12,
  min_points_per_track = 10,
  min_Ninds_per_col = 5
)
}
\arguments{
\item{dataset}{A dataset with location estimates. Must include a species, colony and Month column (tibble)}

\item{species.}{A species' name (character string)}

\item{colony.}{A colony's name (character string)}

\item{month.}{A month (numeric)}

\item{min_points_per_track}{(numeric)}

\item{min_Ninds_per_col}{(numeric)}
}
\value{
A results list with the total filtered dataset and a vector containing the migration ids as strings
}
\description{
Filter for tracks to be assessed in the colony representativeness analysis
}
