% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE_UD_fun.R
\name{KDE_UD_fun}
\alias{KDE_UD_fun}
\title{Determine Utilization Density}
\usage{
KDE_UD_fun(migration_track, h_par)
}
\arguments{
\item{migration_track}{individual-level migration track, with a longitude and latitude column}

\item{h_par}{smoothing parameter (numeric)}
}
\value{
Returns a list with two elements>
1) An object of class `estUD` as produced by `adehabitatHR::kernelUD()`.
2) The track-specific projection (character string)
}
\description{
Determine Utilization Density
}
